feat: Pass requestAgentOptions through to the jwks-rsa library
NON-BREAKING: this adds an optional configuration parameter to the OktaJwtVerifier constructor, which in turn is passed
to the jwksClient constructor. Internally, those options are passed on to Axios, thus giving us the ability to configure the
https agent used to fetch the JWKs, for example if we need to trust a specific certification authority.